UserId check added to deleting and marking lost dogs as found.

+++ Backend/Backend/Controllers/LostDogController.cs
@@ -102,9 +102,19 @@ namespace Backend.Controllers
         [Route("{dogId}/found")]
         public async Task<IActionResult> MarkLostDogAsFound(int dogId)
         {
-            var serviceResponse = await lostDogService.MarkLostDogAsFound(dogId);
-            var controllerResponse = mapper.Map<ServiceResponse, ControllerResponse>(serviceResponse);
-            return StatusCode(serviceResponse.StatusCode, controllerResponse);
+            var savedDog = await lostDogService.GetLostDogDetails(dogId);
+            if (User.Claims.First(c => c.Type == ClaimTypes.NameIdentifier)?.Value == savedDog.Data.OwnerId.ToString())
+            {
+                var serviceResponse = await lostDogService.MarkLostDogAsFound(dogId);
+                var controllerResponse = mapper.Map<ServiceResponse, ControllerResponse>(serviceResponse);
+                return StatusCode(serviceResponse.StatusCode, controllerResponse);
+            }
+            else
+                return Unauthorized(new ControllerResponse()
+                {
+                    Message = "Attempted to mark as found a dog which is not owned by the user!",
+                    Successful = false
+                });
         }
 
 
@@ -112,18 +122,22 @@ namespace Backend.Controllers
         [Route("{dogId}")]
         public async Task<IActionResult> DeleteLostDog(int dogId)
         {
-            var serviceResponse = await lostDogService.DeleteLostDog(dogId);
-            var controllerResponse = mapper.Map<ServiceResponse, ControllerResponse>(serviceResponse);
-            return StatusCode(serviceResponse.StatusCode, controllerResponse);
+            var savedDog = await lostDogService.GetLostDogDetails(dogId);
+            if (User.Claims.First(c => c.Type == ClaimTypes.NameIdentifier)?.Value == savedDog.Data.OwnerId.ToString())
+            {
+                var serviceResponse = await lostDogService.DeleteLostDog(dogId);
+                var controllerResponse = mapper.Map<ServiceResponse, ControllerResponse>(serviceResponse);
+                return StatusCode(serviceResponse.StatusCode, controllerResponse);
+            }
+            else
+                return Unauthorized(new ControllerResponse()
+                {
+                    Message = "Attempted to delete a dog which is not owned by the user!",
+                    Successful = false
+                });
+
         }
 
-        //[HttpGet]
-        //[Route]
-        //public async Task<IActionResult> GetUserLostDogs(int ownerId)
-        //{
-        //    var serviceResponse = await _lostDogService.GetUserLostDogs(ownerId);
-        //    return StatusCode(serviceResponse.StatusCode, serviceResponse);
-        //}
         //[HttpPost]
         //[Route("{}/comment")]
         //public async Task<IActionResult> AddLostDogComment(AddLostDogCommentDto commentDto)
